example-flask-mongoengine/requirements.txt
flask-script
flask-mongoengine
python-social-auth-app-flask-mongoengine
